pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "radiology-ai"
version = "0.1.0"
description = "Production-grade multi-agent AI radiology report generation system"
requires-python = ">=3.11"

dependencies = [
    # --- Agent orchestration ---
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",       # OpenRouter is OpenAI-compatible

    # --- API ---
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",       # file uploads
    "python-jose[cryptography]>=3.3.0", # JWT
    "passlib[bcrypt]>=1.7.4",

    # --- Task queue ---
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # --- Medical imaging ---
    "pydicom>=2.4.0",
    "SimpleITK>=2.3.0",
    "monai>=1.3.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",

    # --- RAG / vector DB ---
    "qdrant-client>=1.9.0",
    "llama-index>=0.10.0",
    "llama-index-vector-stores-qdrant>=0.2.0",

    # --- Database ---
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",        # swap for asyncpg if going fully async

    # --- AWS ---
    "boto3>=1.34.0",
    "botocore>=1.34.0",

    # --- MLOps ---
    "mlflow>=2.13.0",
    "wandb>=0.17.0",

    # --- Observability ---
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",

    # --- UI ---
    "gradio>=4.36.0",

    # --- Compliance / security ---
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",                 # async test client for FastAPI
    "ruff>=0.4.0",                   # linter + formatter
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "pipeline*", "ui*", "mlops*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=70"